solution/src/main/java/com/alanjwilliams/engage/service/ExpenseServiceImpl.java
package com.alanjwilliams.engage.service;

import com.alanjwilliams.engage.entity.Expense;
import com.alanjwilliams.engage.repository.ExpenseRepository;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;

import javax.transaction.Transactional;
import java.util.List;

/**
 * Implementation of the Expense Service working with the Expense Repository.
 */
@Service
@Transactional
public class ExpenseServiceImpl implements ExpenseService {

    @Autowired
    private ExpenseRepository expenseRepository;

    @Override
    public List<Expense> getAll() {
        return expenseRepository.findAll();
    }

    @Override
    public Expense save(Expense expense) {
        return expenseRepository.save(expense);
    }
}
